requirements.txt
requests
python-dotenv
feedparser
gnews

